move parseclustersresponse
